requirements.txt
djangorestframework-simplejwt
djangorestframework
